+++ examples/dataframe.py
@@ -25,3 +25,8 @@ if __name__ == "__main__":
     SQL = "SELECT * FROM mydf LIMIT 2 OFFSET 1"
     for row in cursor.execute(SQL):
         print(row)
+
+    emptydf = pd.DataFrame({"a": []})
+    SQL = "SELECT * from emptydf WHERE a = 'test'"
+    for row in cursor.execute(SQL):
+        print(row)