+++ src/shillelagh/adapters/api/generic_json.py
@@ -6,7 +6,7 @@ An adapter for fetching JSON data.
 
 import logging
 import urllib.parse
-from typing import Any, Dict, Iterator, List, Optional, Tuple
+from typing import Any, Dict, Iterator, List, Optional, Set, Tuple
 
 import requests_cache
 from jsonpath import JSONPath
@@ -22,17 +22,21 @@ _logger = logging.getLogger(__name__)
 
 SUPPORTED_PROTOCOLS = {"http", "https"}
 AVERAGE_NUMBER_OF_ROWS = 100
+CACHE_EXPIRATION = 180
 
 
-def get_session() -> requests_cache.CachedSession:
+def get_session(request_headers: Dict[str, str]) -> requests_cache.CachedSession:
     """
     Return a cached session.
     """
-    return requests_cache.CachedSession(
+    session = requests_cache.CachedSession(
         cache_name="generic_json_cache",
         backend="sqlite",
-        expire_after=180,
+        expire_after=CACHE_EXPIRATION,
     )
+    session.headers.update(request_headers)
+
+    return session
 
 
 class GenericJSONAPI(Adapter):
@@ -45,6 +49,7 @@ class GenericJSONAPI(Adapter):
 
     supports_limit = False
     supports_offset = False
+    supports_requested_columns = True
 
     @staticmethod
     def supports(uri: str, fast: bool = True, **kwargs: Any) -> Optional[bool]:
@@ -54,7 +59,8 @@ class GenericJSONAPI(Adapter):
         if fast:
             return Maybe
 
-        session = get_session()
+        request_headers = kwargs.get("request_headers", {})
+        session = get_session(request_headers)
         response = session.head(uri)
         return "application/json" in response.headers.get("content-type", "")
 
@@ -67,13 +73,18 @@ class GenericJSONAPI(Adapter):
 
         return uri, path
 
-    def __init__(self, uri: str, path: str = "$[*]"):
+    def __init__(
+        self,
+        uri: str,
+        path: str = "$[*]",
+        request_headers: Optional[Dict[str, str]] = None,
+    ):
         super().__init__()
 
         self.uri = uri
         self.path = path
 
-        self._session = get_session()
+        self._session = get_session(request_headers or {})
 
         self._set_columns()
 
@@ -98,12 +109,13 @@ class GenericJSONAPI(Adapter):
 
     get_cost = SimpleCostModel(AVERAGE_NUMBER_OF_ROWS)
 
-    def get_data(  # pylint: disable=unused-argument
+    def get_data(  # pylint: disable=unused-argument, too-many-arguments
         self,
         bounds: Dict[str, Filter],
         order: List[Tuple[str, RequestedOrder]],
         limit: Optional[int] = None,
         offset: Optional[int] = None,
+        requested_columns: Optional[Set[str]] = None,
         **kwargs: Any,
     ) -> Iterator[Row]:
         response = self._session.get(self.uri)
@@ -113,6 +125,11 @@ class GenericJSONAPI(Adapter):
 
         parser = JSONPath(self.path)
         for i, row in enumerate(parser.parse(payload)):
+            row = {
+                k: v
+                for k, v in row.items()
+                if requested_columns is None or k in requested_columns
+            }
             row["rowid"] = i
             _logger.debug(row)
             yield flatten(row)