+++ CHANGELOG.rst
@@ -5,6 +5,15 @@ Changelog
 Next
 ====
 
+- Allow specifying custom request headers when using the generic JSON adapter (#337)
+- Fix for escaping identifiers correctly (#340)
+- Support for S3-compatible storage (#343)
+- Adapters can now know which columns were requested (#345)
+- Python 3.11 officially supported (#334)
+- Fix for error when an adapater can't be loaded (#346)
+- Fix for ``BestIndexObject`` (#350)
+- Fix for empty dataframes (#351)
+
 Version 1.2.0 - 2023-02-17
 ==========================